# lib/api-zod/src/generated/types/createCitationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenJenni Academic AI Writing Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCitationBodyCitationStyle } from "./createCitationBodyCitationStyle";

export interface CreateCitationBody {
  documentId: number;
  author: string;
  year?: string | null;
  title: string;
  journal?: string | null;
  volume?: string | null;
  pages?: string | null;
  doi?: string | null;
  url?: string | null;
  pdfId?: number | null;
  citationStyle: CreateCitationBodyCitationStyle;
}