# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenJenni Academic AI Writing Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentCitationStyle } from "./documentCitationStyle";

export interface Document {
  id: number;
  title: string;
  content: string;
  citationStyle: DocumentCitationStyle;
  wordCount: number;
  targetWordCount?: number | null;
  fieldOfStudy?: string | null;
  createdAt: Date;
  updatedAt: Date;
}