# lib/api-zod/src/generated/types/outlineSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenJenni Academic AI Writing Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface OutlineSection {
  level: number;
  title: string;
  subsections: OutlineSection[];
}