# lib/api-zod/src/generated/types/paraphraseResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenJenni Academic AI Writing Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ParaphraseResult {
  result: string;
  warning?: string | null;
}